# Conflicts:
#	Gelderlandplein/app/src/main/java/com/example/gelderlandplein/ui/MainActivity.kt
#	Gelderlandplein/app/src/main/res/values/strings.xml

+++ Gelderlandplein/app/src/main/java/com/example/gelderlandplein/models/Shop.kt
@@ -11,8 +11,8 @@ data class Shop(
     var tag: String? = "",
     var image: String? = "",
     var openingstijden: ArrayList<String>?,
-    var latitude: Float,
-    var longitude: Float,
+    var latitude: Double,
+    var longitude: Double,
     var inventory: ArrayList<String>?
 
 ) : Parcelable {
